Fix homepage and 404 mobile menu
