+++ website/src/components/index.js
@@ -3,5 +3,6 @@ export { Container } from './Container';
 export { Header } from './Header';
 export { Footer } from './Footer';
 export { Sidebar } from './Sidebar';
+export { SidebarNav } from './SidebarNav';
 export { Search } from './Search';
 export { SiteMeta } from './SiteMeta';
