+++ website/src/templates/docs.js
@@ -1,10 +1,10 @@
 /** @jsx jsx */
 
-import React, { useEffect, useRef, useState } from 'react'; // eslint-disable-line no-unused-vars
+import React from 'react'; // eslint-disable-line no-unused-vars
 import { Helmet } from 'react-helmet';
 import { Link, graphql } from 'gatsby';
 import MDXRenderer from 'gatsby-mdx/mdx-renderer';
-import throttle from 'lodash.throttle';
+
 import { MDXProvider } from '@mdx-js/tag';
 import { jsx } from '@emotion/core';
 import { SkipNavContent } from '@reach/skip-nav';
@@ -19,10 +19,9 @@ import Layout from '../templates/layout';
 import mdComponents from '../components/markdown';
 import { SiteMeta } from '../components/SiteMeta';
 import { media, mediaMax } from '../utils/media';
-import { Container, Footer, Sidebar, Search } from '../components';
+import { Container } from '../components';
 import { CONTAINER_GUTTERS } from '../components/Container';
-
-const SIDEBAR_WIDTH = 280;
+import { Sidebar, SIDEBAR_WIDTH } from '../components/Sidebar';
 
 function titleCase(str, at = '-') {
   if (!str) return str;
@@ -60,11 +59,7 @@ export default function Template({
       <Layout>
         {({ sidebarOffset, sidebarIsVisible }) => (
           <Container>
-            <Aside isVisible={sidebarIsVisible} offsetTop={sidebarOffset} key="sidebar">
-              <Search />
-              <Sidebar />
-              <Footer />
-            </Aside>
+            <Sidebar isVisible={sidebarIsVisible} offsetTop={sidebarOffset} />
             <Content>
               <main>
                 <SkipNavContent />
@@ -252,81 +247,7 @@ const PaginationButton = ({ align = 'left', ...props }) => (
 // ==============================
 
 const layoutGutter = gridSize * 4;
-let oldSidebarOffset = 0;
-let oldWindowOffset = 0;
 
-const Aside = ({ offsetTop, isVisible, ...props }) => {
-  const asideRef = useRef();
-  const [isStuck, setSticky] = useState(false);
-
-  const handleWindowScroll = () => {
-    oldWindowOffset = window.pageYOffset;
-    if (window.pageYOffset > offsetTop && !isStuck) {
-      setSticky(true);
-    }
-    if (window.pageYOffset <= offsetTop && isStuck) {
-      setSticky(false);
-    }
-  };
-
-  const maintainSidebarScroll = throttle(() => {
-    oldSidebarOffset = asideRef.current.scrollTop;
-  });
-
-  useEffect(() => {
-    const asideEl = asideRef.current; // maintain ref for cleanup
-    window.addEventListener('scroll', handleWindowScroll);
-    asideEl.addEventListener('scroll', maintainSidebarScroll);
-
-    // cleanup
-    return () => {
-      window.removeEventListener('scroll', handleWindowScroll);
-      asideEl.removeEventListener('scroll', maintainSidebarScroll);
-    };
-  });
-
-  // NOTE: maintain the user's scroll whilst navigating between pages.
-  // This is a symptom of Gatsby remounting the entire tree (template) on each
-  // page change via `createPage` in "gatsby-node.js".
-  useEffect(() => {
-    const scrollTop = oldWindowOffset ? oldSidebarOffset + offsetTop : oldSidebarOffset;
-    asideRef.current.scrollTop = scrollTop;
-  }, [asideRef.current]);
-
-  const stickyStyles = {
-    height: isStuck ? '100%' : `calc(100% - ${offsetTop}px)`,
-    position: isStuck ? 'fixed' : 'absolute',
-    width: SIDEBAR_WIDTH,
-    top: isStuck ? 0 : offsetTop,
-  };
-
-  // NOTE: the 5px gutter is to stop inner elements outline/box-shadow etc.
-  // being cropped because the aside has overflow-x hidden (due to y=auto).
-  const avoidCropGutter = 5;
-
-  return (
-    <aside
-      ref={asideRef}
-      css={{
-        boxSizing: 'border-box',
-        overflowY: 'auto',
-        paddingBottom: '3rem',
-        paddingTop: layoutGutter,
-        marginLeft: -avoidCropGutter,
-        paddingLeft: avoidCropGutter,
-
-        [mediaMax.sm]: {
-          display: isVisible ? 'block' : 'none',
-        },
-        [media.sm]: {
-          paddingRight: layoutGutter,
-          ...stickyStyles,
-        },
-      }}
-      {...props}
-    />
-  );
-};
 const Content = props => (
   <div
     css={{
